hypervisor: adds hypervisor Driver for RPC
This adds a new hypervisor Driver backed by libvirt RPC.

The most notable change here is that libvirtrpc.Disconnect() now closes
the underlying net.Conn. I've included the change here in order to keep
the driver relatively simple. However, we'll need to make this change
either way. It turns out that when Disconnect() is called, libvirt
closes its half of the connection. Not closing our half provides a
confusing interface for the caller. If the connection were to be reused,
subsequent calls would fail with a TCP RST.

Fixes #42