qmp: expand on new testing style
This modifies `testMonitor.Run()` to allow the closure to return a
`qmp.Monitor.Run()` failure. Maybe it's a bit much to include tests
`TestVersionMonitorFail()`? In the end I think tests that
include failure scenarios will be a good thing to have.

I've also ported a few other tests over to the new style. We'll need
to chop up some of our nested types to make porting the rest easier.