Be more lenient about isolated docstrings in the QMP spec.
Recent revisions of the spec have added more general documentation
(good!), but those are tripping up the code generator as potential
rogue datatype docstrings (bad!).

In practice, we still have a safety in place for that, because we
refuse to codegen a type definition that doesn't have a docstring.
So, we can safely ignore orphan docstrings, as long as we don't
ignore orphan definitions.