Snapshot portability across libvirt 3 and 6
There have been changes to how libvirt creates snapshots, that make the
current go-qemu APIs unable to work on libvirt 6, since they relied
on using a device string that is no longer present.

Fortunately, the call sequence and arguments that work for libvirt 6
*also* work for libvirt 3.

Unfortunately, they require information specifically about the block job
being cancelled, so the API changes slightly.

Block jobs were being completed/cancelled on the BlockDevice,
but the information used to identify them in libvirt6 is not the
BlockDevice.Device, but the BlockJob.Device, so move the methods to the
BlockJob.

See:
- https://www.mail-archive.com/qemu-discuss@nongnu.org/msg05225.html