transition to functional GitHub Actions CI
Travis CI has historically waffled back and forth about how it feels
about a free-tier plan for open source projects. It's a rollercoaster
that I personally don't ever want to ride again.

Furthermore, go-qemu's CI seems to be down in the doldrums, so replace
it with a simple GitHub Actions workflow.