avoid name collisions between variant and struct types
If generating against a QEMU distribution that has unions with base
fields, the code generator would generate structs that conflict with
each other. For example, BlockdevOptions results in code generation with
a number of structs with the same name (i.e., "BlockdevOptionsGluster"),
where you'd see a struct emitted for the variant that contains the base
member fields and also a standalone BlockdevOptionsGluster which is
defined separately in a QAPI document.

Because of how union bases work, they are technically distinct structs,
so make them distinct.